Refactor EnrichedTeam interface and implement TeamCard component for improved team data display
