+++ apps/web/app/briefing/page.tsx
@@ -9,6 +9,9 @@ import { matchLabel, sortMatches } from '../../lib/match-utils';
 import { useSimulationEpa } from '../../hooks/use-simulation-epa';
 import { InfoBox } from '../../components/info-box';
 import { LoadingSpinner } from '../../components/loading-spinner';
+import { TeamCard } from '../../components/team-card';
+import { getAdapter } from '@allianceops/shared';
+import type { EnrichedTeam } from '../../lib/types';
 
 interface TBAMatch {
   key: string;
@@ -22,92 +25,15 @@ interface TBAMatch {
   winning_alliance: string;
 }
 
-interface EnrichedTeam {
-  team_number: number;
-  nickname: string;
-  epa: { total: number; auto: number; teleop: number; endgame: number } | null;
-  eventRecord: { wins: number; losses: number; ties: number } | null;
-  winrate: number | null;
-  tbaRank: number | null;
-}
-
-function teamNum(key: string): string {
-  return key.replace('frc', '');
-}
-
-function EpaBar({ value, max, color }: { value: number; max: number; color: string }) {
-  const pct = Math.min((value / max) * 100, 100);
-  return (
-    <div className="w-full bg-gray-200 dark:bg-gray-700 rounded-full h-2">
-      <div className={`h-2 rounded-full ${color}`} style={{ width: `${pct}%` }} />
-    </div>
-  );
-}
-
-function TeamCard({
-  teamKey,
-  epaMap,
-  record,
-  epaRank,
-}: {
-  teamKey: string;
-  epaMap: Map<number, EnrichedTeam>;
-  record?: { wins: number; losses: number; ties: number };
-  epaRank?: number | null;
-}) {
-  const num = parseInt(teamKey.replace('frc', ''), 10);
-  const data = epaMap.get(num);
-  const maxEpa = 40;
-  const displayRecord = record ?? data?.eventRecord;
-
-  return (
-    <div className="rounded-lg border border-gray-200 dark:border-gray-700 p-3 space-y-2">
-      <div className="flex justify-between items-center">
-        <span className="font-bold text-lg">{teamNum(teamKey)}</span>
-        <div className="flex items-center gap-2 text-xs text-gray-500">
-          {data?.tbaRank != null && <span>Rank #{data.tbaRank}</span>}
-          {epaRank != null && <span>EPA #{epaRank}</span>}
-          {displayRecord && (
-            <span>
-              {displayRecord.wins}W-{displayRecord.losses}L
-            </span>
-          )}
-        </div>
-      </div>
-      {data?.epa?.total != null ? (
-        <div className="space-y-1 text-xs">
-          <div className="flex items-center gap-2">
-            <span className="w-14">Total</span>
-            <EpaBar value={data.epa.total} max={maxEpa} color="bg-primary-500" />
-            <span className="w-8 text-right">{data.epa.total.toFixed(1)}</span>
-          </div>
-          <div className="flex items-center gap-2">
-            <span className="w-14">Auto</span>
-            <EpaBar value={data.epa.auto ?? 0} max={maxEpa / 2} color="bg-green-500" />
-            <span className="w-8 text-right">{(data.epa.auto ?? 0).toFixed(1)}</span>
-          </div>
-          <div className="flex items-center gap-2">
-            <span className="w-14">Teleop</span>
-            <EpaBar value={data.epa.teleop ?? 0} max={maxEpa / 2} color="bg-blue-500" />
-            <span className="w-8 text-right">{(data.epa.teleop ?? 0).toFixed(1)}</span>
-          </div>
-          <div className="flex items-center gap-2">
-            <span className="w-14">Endgame</span>
-            <EpaBar value={data.epa.endgame ?? 0} max={maxEpa / 3} color="bg-purple-500" />
-            <span className="w-8 text-right">{(data.epa.endgame ?? 0).toFixed(1)}</span>
-          </div>
-        </div>
-      ) : (
-        <p className="text-xs text-gray-400">No EPA data</p>
-      )}
-    </div>
-  );
-}
-
 export default function BriefingPage() {
   const { eventKey, teamNumber, year } = useEventSetup();
   const { activeCursor } = useSimulation();
   const myTeamKey = `frc${teamNumber}`;
+
+  const adapter = getAdapter(year);
+  const cardMetrics = (adapter?.gameSpecificMetrics ?? []).filter(
+    (m) => m.renderLocation === 'team_card' || m.renderLocation === 'all',
+  );
 
   const { data: rawMatches, loading: matchesLoading } = useApi<TBAMatch[]>(
     eventKey ? `event/${eventKey}/matches` : null,
@@ -249,6 +175,7 @@ export default function BriefingPage() {
                 key={t}
                 teamKey={t}
                 epaMap={epaMap}
+                metrics={cardMetrics}
                 epaRank={epaRankMap.get(parseInt(t.replace('frc', ''), 10))}
                 record={activeCursor !== null && matches ? getTeamRecord(matches, t, activeCursor) : undefined}
               />
@@ -265,6 +192,7 @@ export default function BriefingPage() {
                 key={t}
                 teamKey={t}
                 epaMap={epaMap}
+                metrics={cardMetrics}
                 epaRank={epaRankMap.get(parseInt(t.replace('frc', ''), 10))}
                 record={activeCursor !== null && matches ? getTeamRecord(matches, t, activeCursor) : undefined}
               />
